pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "releaf-ai"
version = "0.1.0"
description = "AI-powered sustainability and waste intelligence platform"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "ReleAF AI Team"}
]
keywords = ["ai", "sustainability", "recycling", "upcycling", "computer-vision", "nlp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core ML frameworks
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.35.0",
    "peft>=0.7.0",
    "accelerate>=0.25.0",
    "bitsandbytes>=0.41.0",
    
    # Vision
    "timm>=0.9.0",
    "ultralytics>=8.0.0",
    "opencv-python>=4.8.0",
    "albumentations>=1.3.0",
    "pillow>=10.0.0",
    
    # NLP & Embeddings
    "sentence-transformers>=2.2.0",
    "tokenizers>=0.15.0",
    
    # Vector stores & RAG
    "qdrant-client>=1.7.0",
    "faiss-cpu>=1.7.4",
    "chromadb>=0.4.0",
    "rank-bm25>=0.2.2",
    
    # Graph
    "neo4j>=5.14.0",
    "torch-geometric>=2.4.0",
    "networkx>=3.2.0",
    
    # Web & API
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.12.0",
    
    # Data processing
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "pyarrow>=14.0.0",
    "datasets>=2.15.0",
    
    # Utilities
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "loguru>=0.7.0",
    
    # Monitoring
    "prometheus-client>=0.19.0",
    "wandb>=0.16.0",
    
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

gpu = [
    "faiss-gpu>=1.7.4",
]

all = [
    "releaf-ai[dev,gpu]",
]

[project.urls]
Homepage = "https://github.com/releaf-ai/sustainability-ai-model"
Documentation = "https://docs.releaf-ai.com"
Repository = "https://github.com/releaf-ai/sustainability-ai-model"
Issues = "https://github.com/releaf-ai/sustainability-ai-model/issues"

[tool.setuptools]
packages = ["services", "training", "scripts"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | models
  | data
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=services --cov=training --cov-report=html --cov-report=term"

